Use long for nzbx rageid

+++ NzbDrone.Core/Indexers/Nzbx/NzbxRecentItem.cs
@@ -36,7 +36,7 @@ namespace NzbDrone.Core.Indexers.Nzbx
         public int CategoryId { get; set; }
         public string ImdbId { get; set; }
         public string AnidbId { get; set; }
-        public int RageId { get; set; }
+        public long RageId { get; set; }
         public int Comments { get; set; }
         public int Downloads { get; set; }
     }
